Add cdbuf.vim
Adds a CD command to :cd to the PWD to the directory of the current
buffer. Can be set to only use :lcd if g:cdbuf_global is set to 0. By
default it will just use :cd.

CD - uses :[l]cd %:p:h to change the PWD, respecting g:cdbuf_global.
CD! - same as :CD, but ignores g:cdbuf_global.