Add unite-ack.vim - ack/ag for unite.
If ack or ag is available, one of the two is used. ag is preferred if
both are available.

If ag is used, the --skip-vcs-ignores flag is passed to ag.

Pretty sure this file was based off someone else's vimrc, but I can't
remember whose and it's kind of barebones/simple, so I'm not inclined to
go find it.