Remove nerdtree/bindings. Add vim-vinegar.
Kind of prefer just having the directory tree open in the current
window rather than pop up on the side. Vinegar just adds some stuff
I like, but it could probably be removed too in the future.

Added <Space>e as an equivalent of ,sd because the latter is a little
harder to type (the comma is slighlty out of the way given how I use
a keyboard).